requirements.txt
fastapi
uvicorn
python-dotenv
langchain
chromadb
openai
langchain-community
pypdf
jq
tiktoken